timing.py: keep parameters as float128 values in tm_delay

+++ enterprise_extensions/timing.py
@@ -31,7 +31,7 @@ def tm_delay(residuals, t2pulsar, tmparams_orig, tmparams, which='all'):
     orig_params = np.array([tmparams_orig[key] for key in keys])
 
     # put varying parameters into dictionary
-    tmparams_rescaled = np.atleast_1d(np.double(orig_params[:, 0] +
+    tmparams_rescaled = np.atleast_1d(np.float128(orig_params[:, 0] +
                                                 tmparams * orig_params[:, 1]))
     tmparams_vary = OrderedDict(zip(keys, tmparams_rescaled))
 
@@ -41,7 +41,7 @@ def tm_delay(residuals, t2pulsar, tmparams_orig, tmparams, which='all'):
 
     # remember to set values back to originals
     t2pulsar.vals(OrderedDict(zip(keys,
-                                  np.atleast_1d(np.double(orig_params[:, 0])))))
+                                  np.atleast_1d(np.float128(orig_params[:, 0])))))
 
     # Sort the residuals
     isort = np.argsort(t2pulsar.toas(), kind='mergesort')
